refactor: improve code smell for postprocessing
