+++ plugins/plugin-chart-echarts/src/BoxPlot/buildQuery.ts
@@ -16,31 +16,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { buildQueryContext, getMetricLabel } from '@superset-ui/core';
-import { BoxPlotQueryFormData, BoxPlotQueryObjectWhiskerType } from './types';
-
-const PERCENTILE_REGEX = /(\d+)\/(\d+) percentiles/;
+import { buildQueryContext } from '@superset-ui/core';
+import { boxplotOperator } from '@superset-ui/chart-controls';
+import { BoxPlotQueryFormData } from './types';
 
 export default function buildQuery(formData: BoxPlotQueryFormData) {
-  const { columns = [], granularity_sqla, groupby = [], whiskerOptions } = formData;
+  const { columns = [], granularity_sqla, groupby = [] } = formData;
   return buildQueryContext(formData, baseQueryObject => {
-    let whiskerType: BoxPlotQueryObjectWhiskerType;
-    let percentiles: [number, number] | undefined;
-    const { metrics = [] } = baseQueryObject;
-    const percentileMatch = PERCENTILE_REGEX.exec(whiskerOptions as string);
-
-    if (whiskerOptions === 'Tukey' || !whiskerOptions) {
-      whiskerType = 'tukey';
-    } else if (whiskerOptions === 'Min/max (no outliers)') {
-      whiskerType = 'min/max';
-    } else if (percentileMatch) {
-      whiskerType = 'percentile';
-      percentiles = [parseInt(percentileMatch[1], 10), parseInt(percentileMatch[2], 10)];
-    } else {
-      throw new Error(`Unsupported whisker type: ${whiskerOptions}`);
-    }
     const distributionColumns: string[] = [];
-
     // For now default to using the temporal column as distribution column.
     // In the future this control should be made mandatory.
     if (!columns.length && granularity_sqla) {
@@ -51,17 +34,7 @@ export default function buildQuery(formData: BoxPlotQueryFormData) {
         ...baseQueryObject,
         columns: [...distributionColumns, ...columns, ...groupby],
         series_columns: groupby,
-        post_processing: [
-          {
-            operation: 'boxplot',
-            options: {
-              whisker_type: whiskerType,
-              percentiles,
-              groupby,
-              metrics: metrics.map(getMetricLabel),
-            },
-          },
-        ],
+        post_processing: [boxplotOperator(formData, baseQueryObject)],
       },
     ];
   });
