+++ plugins/plugin-chart-echarts/src/Timeseries/buildQuery.ts
@@ -24,6 +24,8 @@ import {
   sortOperator,
   pivotOperator,
   resampleOperator,
+  contributionOperator,
+  prophetOperator,
 } from '@superset-ui/chart-controls';
 
 export default function buildQuery(formData: QueryFormData) {
@@ -40,27 +42,8 @@ export default function buildQuery(formData: QueryFormData) {
         sortOperator(formData, { ...baseQueryObject, is_timeseries: true }),
         rollingWindowOperator(formData, baseQueryObject),
         pivotOperator(formData, { ...baseQueryObject, is_timeseries: true }),
-        formData.contributionMode
-          ? {
-              operation: 'contribution',
-              options: {
-                orientation: formData.contributionMode,
-              },
-            }
-          : undefined,
-        formData.forecastEnabled
-          ? {
-              operation: 'prophet',
-              options: {
-                time_grain: formData.time_grain_sqla,
-                periods: parseInt(formData.forecastPeriods, 10),
-                confidence_interval: parseFloat(formData.forecastInterval),
-                yearly_seasonality: formData.forecastSeasonalityYearly,
-                weekly_seasonality: formData.forecastSeasonalityWeekly,
-                daily_seasonality: formData.forecastSeasonalityDaily,
-              },
-            }
-          : undefined,
+        contributionOperator(formData, baseQueryObject),
+        prophetOperator(formData, baseQueryObject),
       ],
     },
   ]);
